Separate output selection to signal setup.
On startup, the program automatically selects a midi output port.
It also set up a Qt signal so that when the user clicks on the output
ports menu, the menu gets filles up with the available midi output
entries.
These are two - different - things. This commit just makes it now
explicit in the codebase.