Just change key colours, don't redraw the whole keyboard
Now, the gui won't clear the keyboard and recreate the whole 88 keys
on every key event. Instead the keyboard is created only once at the
initialisation, and for each key that get pressed or released, the gui
will only change the colour or that key, without doing anything to any
of the other keys.