Performance improvement, quit the 'detect new signal' asap.
This function is called at each event loop, therefore let's keep its
code execution as short as possible. A further optimisation here would
be to only use one sig_atomic_t variable for all signals, and use a
different number for each of them to know which signal was received.