Bugfix: now plays the measure with last number when playing subsequences
The line
 auto it = std::next(events_begin, static_cast<decltype(events_begin)::difference_type>(stop_pos));
sets the iterator to the beginning of the measure with the last_measure
bar number change. The function had to look for the next event with a
bar_number_change. However it did not increment the iterator before
doing the while loop. As a consequence, the while loop would never
enter (because the iterator is already at an event with a bar number
change), and thus the iterator wouldn't be changed.

In short the function get_end_measure_sequence_pos wasn't doing
anything more than just returning the result of get_starting_measure_event_pos