Make it difficult for the user to connect an output midi port to an input one
Listening to the output produced by the software leads to a kind of
perpetual movement leading to a 100% CPU use. This commit hides the output
port from this program to appear in the list of possible midi ports to
listen to.
It also prevents a port to which this program outputs to, to appear on the
list on inputs, or from a port this program listens to from appearing on
the output menu.
This is to prevent the user from creating this perpetual movement by
mistake. If the user really wants to do so (which is more than likely a
problem), he would then have to use a second program to be able to create
the loop.
This program won't try to detect if there is such a loop, and thus doesn't
implement any smarter mechanism to detect this issue.